chore: update version of node packages
* Update jest configuration
* Update storybook configuration
* Remove/Replace deprecated plugins
