+++ .eslintrc.js
@@ -32,7 +32,7 @@ module.exports = {
     DEVELOPMENT: "readonly",
     DOCKER: "readonly",
   },
-  parser: "babel-eslint",
+  parser: "@babel/eslint-parser",
   parserOptions: {
     ecmaFeatures: {
       jsx: true,
