--- .storybook/addons.ts
@@ -1,8 +1,0 @@
-import "@storybook/addons"
-import "storybook-readme/register"
-import "storybook-addon-styled-component-theme/dist/register"
-import "@storybook/addon-actions/register"
-import "@storybook/addon-links/register"
-import "@storybook/addon-knobs/register"
-import "@storybook/addon-storysource/register"
-import "@storybook/addon-jest/register"
