+++ package.json
@@ -21,105 +21,100 @@
     "cp-dashboard": "cp -R ./lib ../dashboard/node_modules/@netdata/netdata-ui/",
     "to-dashboard": "yarn ts-bundle && yarn cp-dashboard",
     "prepare": "install-peers && yarn ts-bundle",
-    "resolveAlias": "tscpaths -p tsconfig.json -s ./src -o ./lib"
+    "resolveAlias": "tsconfig-replace-paths -p tsconfig.json -s ./src -o ./lib"
   },
   "keywords": [
     "uikit",
+    "ui-kit",
     "netdata-ui"
   ],
   "author": "netdata",
   "license": "ISC",
   "dependencies": {
-    "@elastic/react-search-ui": "^1.5.1",
-    "@elastic/search-ui-site-search-connector": "^1.5.1",
+    "@elastic/react-search-ui": "^1.14.0",
+    "@elastic/search-ui-site-search-connector": "^1.14.0",
     "@netdata/react-filter-box": "^1.0.0",
-    "@prismicio/client": "^5.1.0",
+    "@prismicio/client": "^6.6.1",
     "@styled-system/css": "^5.1.5",
-    "@tanstack/table-core": "^8.0.0",
-    "@testing-library/user-event": "^13.5.0",
-    "axios": "^0.21.1",
-    "polished": "^4.0.3",
-    "ramda": "^0.27.1",
-    "react-beautiful-dnd": "^13.0.0",
-    "react-portal": "^4.2.0",
-    "react-select": "^5.3.2",
-    "react-table": "^7.2.1",
-    "react-use": "^15.3.4",
-    "react-window": "^1.8.5",
+    "@tanstack/table-core": "8.0.0",
+    "axios": "^0.27.2",
+    "polished": "^4.2.2",
+    "ramda": "^0.28.0",
+    "react-beautiful-dnd": "^13.1.0",
+    "react-portal": "^4.2.2",
+    "react-select": "^5.4.0",
+    "react-table": "^7.8.0",
+    "react-use": "^17.4.0",
+    "react-window": "^1.8.7",
     "styled-system": "^5.1.5",
-    "throttle-debounce": "^3.0.1"
+    "throttle-debounce": "^5.0.0"
   },
   "peerDependencies": {
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "styled-components": "^5.2.1"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "styled-components": "^5.3.5"
   },
   "repository": {
     "type": "git",
     "url": "git@github.com:netdata/netdata-ui.git"
   },
   "devDependencies": {
-    "@babel/plugin-transform-runtime": "^7.12.1",
-    "@babel/preset-env": "^7.12.1",
-    "@babel/preset-react": "^7.12.5",
-    "@netdata/eslint-config-netdata": "0.0.1",
-    "@storybook/addon-actions": "^6.1.0",
-    "@storybook/addon-centered": "^5.2.3",
-    "@storybook/addon-jest": "^6.1.0",
-    "@storybook/addon-knobs": "^6.1.0",
-    "@storybook/addon-links": "^6.1.0",
-    "@storybook/addon-storysource": "^6.1.0",
-    "@storybook/addons": "^6.1.0",
-    "@storybook/react": "^6.1.0",
-    "@testing-library/jest-dom": "^5.11.6",
-    "@testing-library/react": "^11.2.1",
-    "@testing-library/react-hooks": "^3.7.0",
-    "@types/codemirror": "0.0.98",
-    "@types/jest": "^26.0.15",
-    "@types/node": "^14.14.8",
-    "@types/pegjs": "^0.10.2",
-    "@types/ramda": "^0.27.32",
-    "@types/react": "^16.9.5",
-    "@types/react-portal": "^4.0.2",
-    "@types/storybook__react": "^5.2.1",
-    "@types/styled-components": "^5.1.2",
+    "@babel/core": "^7.18.6",
+    "@babel/eslint-parser": "^7.18.2",
+    "@babel/plugin-transform-runtime": "^7.18.6",
+    "@babel/preset-env": "^7.18.6",
+    "@babel/preset-react": "^7.18.6",
+    "@netdata/eslint-config-netdata": "^0.1.2",
+    "@storybook/addon-actions": "^6.5.9",
+    "@storybook/addon-centered": "^5.3.21",
+    "@storybook/addon-essentials": "^6.5.9",
+    "@storybook/addon-interactions": "^6.5.9",
+    "@storybook/addon-jest": "^6.5.9",
+    "@storybook/addon-knobs": "^6.4.0",
+    "@storybook/addon-links": "^6.5.9",
+    "@storybook/addon-storysource": "^6.5.9",
+    "@storybook/builder-webpack4": "^6.5.9",
+    "@storybook/builder-webpack5": "^6.5.9",
+    "@storybook/manager-webpack4": "^6.5.9",
+    "@storybook/manager-webpack5": "^6.5.9",
+    "@storybook/react": "^6.5.9",
+    "@storybook/testing-library": "^0.0.13",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/react-hooks": "^8.0.1",
+    "@testing-library/user-event": "^14.2.1",
+    "@types/jest": "^28.1.4",
+    "@types/node": "^18.0.3",
+    "@types/ramda": "^0.28.14",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
+    "@types/react-portal": "^4.0.4",
+    "@types/styled-components": "^5.1.25",
     "@types/styled-system": "^5.1.15",
     "@types/styled-system__css": "^5.0.16",
-    "@typescript-eslint/eslint-plugin": "^4.8.1",
-    "@typescript-eslint/parser": "^4.8.1",
-    "awesome-typescript-loader": "^5.2.1",
-    "babel-eslint": "^10.1.0",
-    "babel-loader": "^8.0.6",
+    "@typescript-eslint/eslint-plugin": "^5.30.5",
+    "@typescript-eslint/parser": "^5.30.5",
+    "babel-loader": "^8.2.5",
     "cpx": "^1.5.0",
-    "eslint": "^7.13.0",
-    "eslint-plugin-react": "^7.22.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
+    "eslint": "^8.19.0",
+    "eslint-plugin-react": "^7.30.1",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "html-loader": "^3.1.2",
     "identity-obj-proxy": "^3.0.0",
-    "install-peers-cli": "^2.1.1",
-    "jest": "^26.6.3",
-    "jest-styled-components": "^7.0.3",
-    "jsdom": "^16.4.0",
-    "json-loader": "^0.5.7",
-    "markdown-loader": "^6.0.0",
-    "prettier": "^2.1.2",
-    "raw-loader": "^4.0.2",
-    "react-redux": "^7.1.1",
-    "react-test-render": "^1.1.2",
-    "react-test-renderer": "^17.0.1",
-    "redux": "^4.0.4",
-    "redux-mock-store": "^1.5.3",
-    "require-context": "^1.1.0",
-    "rimraf": "^3.0.0",
-    "storybook": "^6.1.0",
-    "storybook-addon-styled-component-theme": "^1.2.5",
-    "storybook-readme": "^5.0.8",
-    "styled-components": "^5.2.1",
-    "svg-sprite-loader": "^5.0.0",
-    "svgo": "^1.3.0",
-    "svgo-loader": "^2.2.1",
-    "ts-jest": "^26.4.4",
-    "tscpaths": "^0.0.9",
-    "typescript": "^4.0.5",
+    "install-peers-cli": "^2.2.0",
+    "jest": "^28.1.2",
+    "jest-environment-jsdom": "^28.1.2",
+    "jest-styled-components": "^7.0.8",
+    "jsdom": "19.0.0",
+    "markdown-loader": "^8.0.0",
+    "prettier": "^2.7.1",
+    "rimraf": "^3.0.2",
+    "storybook-readme": "^5.0.9",
+    "svg-sprite-loader": "^6.0.11",
+    "svgo-loader": "^3.0.1",
+    "ts-jest": "^28.0.5",
+    "tsconfig-replace-paths": "^0.0.11",
+    "typescript": "^4.7.4",
     "url-loader": "^4.1.1"
   }
 }
